# source/dotnet/Auth/Clouded.Auth.Provider/auth-swaggerfile.yml
openapi: 3.0.1
info:
  title: Clouded Auth Provider API
  description: Clouded Auth Provider API
  termsOfService: https://clouded.com/terms
  contact:
    name: Email
    email: support@clouded.com
  version: v1
servers:
  - url: /
paths:
...
  /v1/machines/paginated:
    get:
      tags:
        - Machine
      summary: Get all entities paginated
      parameters:
        - name: search
          in: query
          description: ''
          schema:
            type: string
        - name: roleIds
          in: query
          description: ''
          schema:
            type: array
            items:
              type: string
        - name: permissionIds
          in: query
          description: ''
          schema:
            type: array
            items:
              type: string
        - name: page
          in: query
          description: ''
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          description: ''
          schema:
            type: integer
            format: int32
            default: 50
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineDictionaryPaginatedCloudedOkResult'
  '/v1/machines/by_identity/{identity}':
    get:
      tags:
        - Machine
      summary: Get entity by identity
      parameters:
        - name: identity
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: withRoles
          in: query
          description: ''
          schema:
            type: boolean
        - name: withPermissions
          in: query
          description: ''
          schema:
            type: boolean
        - name: withDomains
          in: query
          description: ''
          schema:
            type: boolean
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineDictionaryCloudedOkResult'
  '/v1/machines/{id}':
    get:
      tags:
        - Machine
      summary: Get entity by id
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: withRoles
          in: query
          description: ''
          schema:
            type: boolean
        - name: withPermissions
          in: query
          description: ''
          schema:
            type: boolean
        - name: withDomains
          in: query
          description: ''
          schema:
            type: boolean
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineDictionaryCloudedOkResult'
    patch:
      tags:
        - Machine
      summary: Update entity
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              type: object
              additionalProperties: { }
          text/json:
            schema:
              type: object
              additionalProperties: { }
          application/*+json:
            schema:
              type: object
              additionalProperties: { }
        required: true
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineDictionaryCloudedOkResult'
    delete:
      tags:
        - Machine
      summary: Delete entity
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
  '/v1/machines/{id}/roles/add':
    put:
      tags:
        - Machine
      summary: ''
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MachineRoleInput'
          text/json:
            schema:
              $ref: '#/components/schemas/MachineRoleInput'
          application/*+json:
            schema:
              $ref: '#/components/schemas/MachineRoleInput'
        required: true
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
  '/v1/machines/{id}/roles/remove':
    put:
      tags:
        - Machine
      summary: ''
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MachineRoleInput'
          text/json:
            schema:
              $ref: '#/components/schemas/MachineRoleInput'
          application/*+json:
            schema:
              $ref: '#/components/schemas/MachineRoleInput'
        required: true
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
  '/v1/machines/{id}/permissions/add':
    put:
      tags:
        - Machine
      summary: ''
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MachinePermissionInput'
          text/json:
            schema:
              $ref: '#/components/schemas/MachinePermissionInput'
          application/*+json:
            schema:
              $ref: '#/components/schemas/MachinePermissionInput'
        required: true
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
  '/v1/machines/{id}/permissions/remove':
    put:
      tags:
        - Machine
      summary: ''
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MachinePermissionInput'
          text/json:
            schema:
              $ref: '#/components/schemas/MachinePermissionInput'
          application/*+json:
            schema:
              $ref: '#/components/schemas/MachinePermissionInput'
        required: true
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
  '/v1/machines/{id}/domains/add':
    put:
      tags:
        - Machine
      summary: ''
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MachineDomainInput'
          text/json:
            schema:
              $ref: '#/components/schemas/MachineDomainInput'
          application/*+json:
            schema:
              $ref: '#/components/schemas/MachineDomainInput'
        required: true
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
  '/v1/machines/{id}/domains/remove':
    put:
      tags:
        - Machine
      summary: ''
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MachineDomainInput'
          text/json:
            schema:
              $ref: '#/components/schemas/MachineDomainInput'
          application/*+json:
            schema:
              $ref: '#/components/schemas/MachineDomainInput'
        required: true
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
...
components:
  schemas:
    CloudedExceptionResult:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ExceptionData'
          nullable: true
      additionalProperties: false
...
    ExceptionData:
      required:
        - error
        - i18N
        - type
      type: object
      properties:
        i18N:
          type: string
        error:
          type: string
        type:
          type: string
      additionalProperties: false
...
    MachineDictionaryCloudedOkResult:
      type: object
      properties:
        data:
          type: object
          additionalProperties: { }
          nullable: true
      additionalProperties: false
...
    MachineDictionaryPaginated:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: { }
          description: Paginated data items
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
      additionalProperties: false
    MachineDictionaryPaginatedCloudedOkResult:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/MachineDictionaryPaginated'
      additionalProperties: false
    MachineDomainInput:
      type: object
      properties:
        domainIds:
          type: array
          items: { }
      additionalProperties: false
    MachinePermissionInput:
      type: object
      properties:
        permissionIds:
          type: array
          items: { }
      additionalProperties: false
    MachineRoleInput:
      type: object
      properties:
        roleIds:
          type: array
          items: { }
      additionalProperties: false
...
    PageInfo:
      type: object
      properties:
        pageIndex:
          type: integer
          format: int64
        pageSize:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
      additionalProperties: false
...
  securitySchemes:
    CloudedKey:
      type: apiKey
      description: Enter the `X-CLOUDED-KEY`
      name: X-CLOUDED-KEY
      in: header
    MachineKey:
      type: apiKey
      description: 'Enter the `X-CLOUDED-MACHINE-KEY` base64 string from Machine api and secret key as following: `[API_KEY]:[SECRET_KEY]`'
      name: X-CLOUDED-MACHINE-KEY
      in: header
    Bearer:
      type: apiKey
      description: 'Enter the Bearer Authorization string as following: `Bearer [TOKEN]`'
      name: Authorization
      in: header
security:
  - CloudedKey: [ ]
  - MachineKey: [ ]
  - Bearer: [ ]